% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.NA.R
\name{count.NA}
\alias{count.NA}
\title{Count the missing values (NA) of a variable in a dataset}
\usage{
count.NA(data, col_name)
}
\arguments{
\item{data}{Dataset to retrieve the specific column from}

\item{col_name}{Name of the column (string input) to count missing value(s)}
}
\value{
A tibble (of the dimensions 1 x 1) with variable "col_name" and its value of the integer count of missing values (NA)
}
\description{
\code{count.NA} returns the count of missing values (NA) in a specified column \code{col_name} of a given dataset \code{data}
}
